dataout/publish/kaggle_oilspill_sar_tiles_800x600/raster/ml_pipeline/requirements.txt
numpy>=1.24
pillow>=10.0
tqdm>=4.66
torch>=2.2
torchvision>=0.17
